We have showCometNames & showComets. Read descrip.
I've once tried to add another option such as that existing for planets ("showPlanetImages") but mine would look like "showCometImages".
Still can't figure out if it would be useful since in cometscomponent.cpp we only call showComets. It also addresses to a list of components which makes it different from planetscomponent where each planet is considered an individual object.